Moved FileReader & Fixed Bug in Sales Order Total
Moved file reader objects to own module. Fixed bug that prevented the
sales order total from printing.
